Moving logic to specific component
